pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ecoroute"
description = "EcoRoute Atlas API"
readme = "README.md"
requires-python = ">=3.14"
dynamic = ["version"]
authors = [
    {name = "Shahadath Hossen Sajib", email = "shahadathhossensajib732@gmail.com"}
]
maintainers = [
    {name = "Shahadath Hossen Sajib", email = "shahadathhossensajib732@gmail.com"}
]

dependencies = [
    "bcrypt>=4.2.0",
    "email-validator>=2.3.0",
    "fastapi>=0.136.1",
    "uvicorn[standard]==0.46.0",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "sqlalchemy[asyncio]==2.0.49",
    "asyncpg>=0.31.0",
    "alembic>=1.18.4",
    "greenlet>=3.4.0",
    "qdrant-client>=1.17.1",
    "python-jose[cryptography]>=3.5.0",
    "python-dotenv>=1.2.2",
    "loguru>=0.7.3",
    "python-json-logger>=4.1.0",
    "httpx>=0.28.1",
    "click>=8.1.8",
]

[tool.hatch.version]
path = "app/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "build>=1.4.3",
    "bandit>=1.9.4",
    "python-semantic-release>=10.5.3",
    "ruff>=0.15.12",
    "black>=26.3.1",
    "mypy>=1.20.2",
]

[tool.semantic_release]
version_source = "tag"
version_variables = [
    "app/__init__.py:__version__",
]
branch = "main"
changelog_file = "CHANGELOG.md"
tag_format = "v{version}"
build_command = "uv run python -m build"
upload_to_pypi = false
remove_dist = false
commit_author = "github-actions <actions@github.com>"
commit_parser = "conventional"
